refactor : move SearchController to ShopController
